KS-1-4 refactor variables(seconds)

+++ src/main/kotlin/lesson_1/lesson1task4.kt
@@ -2,12 +2,12 @@ package org.example.lesson_1
 
 fun main(){
 
-    val length:Long = 40_868_600_000
-    val age:UByte = 27u
-    val partOfDay:Float = 0.075f
-    val seconds:Int = 6480
-    val partOfYear:Double = 2.0547945205479453E-4
-    val apogee:Int = 327000
+    val length: Long = 40_868_600_000
+    val age: UByte = 27u
+    val partOfDay: Float = 0.075f
+    val seconds: Short = 6480
+    val partOfYear: Double = 2.0547945205479453E-4
+    val apogee: Int = 327000
 
     println("Расстояние " + length)
     println("Возраст " + age)
